requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiohttp==3.9.1
requests==2.31.0
fake-useragent==1.4.0
python-dotenv==1.0.0
loguru==0.7.2